Make systemd optional
Systemd is used for logging purposes and autostarting the service. For
logging, the code actually already has a fallback for printing to
stdout/stderr which will work fine for systems without systemd. It is
easy to redirect that output to syslog or wherever desired. As for
autostarting the server on device connection, non-systemd systems would
need some kind of dynamic event logic to handle it. That would
ultimately be up to the service manager to implement, but just running
the service manually and/or adding to the initial startup services works
fine for most usecases. So just guard the systemd stuff so it can
compile. Fixes #4.